+++ tests/TestMain.cpp
@@ -22,6 +22,8 @@ int runStateMachineSerializerSerializationTest(int argc, char** argv) __attribut
 int runStateMachineSerializerDeserializationTest(int argc, char** argv) __attribute__((weak));
 int runHsmGraphicsViewTest(int argc, char** argv) __attribute__((weak));
 int runHsmElementDragTest(int argc, char** argv) __attribute__((weak));
+int runAutoGroupItemCursorTest(int argc, char** argv) __attribute__((weak));
+int runHsmTransitionGripVisibilityTest(int argc, char** argv) __attribute__((weak));
 
 int main(int argc, char** argv) {
     qputenv("QT_LOGGING_RULES", "*.debug=true;*.warning=false;");
@@ -56,6 +58,8 @@ int main(int argc, char** argv) {
         runStateMachineSerializerDeserializationTest,
         runHsmGraphicsViewTest,
         runHsmElementDragTest,
+        runAutoGroupItemCursorTest,
+        runHsmTransitionGripVisibilityTest,
     };
 
     for (const auto test : tests) {